make paragraph_db_input_creator flexible and up-to-date
